added swagger doc comments for resource handler endpoints
