+++ resource-service/handler/project_handler.go
@@ -2,7 +2,6 @@ package handler
 
 import (
 	"github.com/gin-gonic/gin"
-	"github.com/keptn/keptn/resource-service/common"
 )
 
 type IProjectHandler interface {
@@ -13,24 +12,58 @@ type IProjectHandler interface {
 
 type ProjectHandler struct {
 	ProjectManager IProjectManager
-	EventSender    common.EventSender
 }
 
-func NewProjectHandler(projectManager IProjectManager, eventSender common.EventSender) *ProjectHandler {
+func NewProjectHandler(projectManager IProjectManager) *ProjectHandler {
 	return &ProjectHandler{
 		ProjectManager: projectManager,
-		EventSender:    eventSender,
 	}
 }
 
+// CreateProject godoc
+// @Summary Create a new project
+// @Description Create a new project
+// @Tags Project
+// @Security ApiKeyAuth
+// @Accept  json
+// @Produce  json
+// @Param   project     body    models.CreateProjectParams     true        "Project"
+// @Success 204 {string} "ok"
+// @Failure 400 {object} models.Error "Invalid payload"
+// @Failure 500 {object} models.Error "Internal error"
+// @Router /project [post]
 func (ph *ProjectHandler) CreateProject(c *gin.Context) {
 
 }
 
+// UpdateProject godoc
+// @Summary Updates an existing project
+// @Description Updates an existing project
+// @Tags Project
+// @Security ApiKeyAuth
+// @Accept  json
+// @Produce  json
+// @Param   project     body    models.UpdateProjectParams     true        "Project"
+// @Success 204 {string} "ok"
+// @Failure 400 {object} models.Error "Invalid payload"
+// @Failure 500 {object} models.Error "Internal error"
+// @Router /project [put]
 func (ph *ProjectHandler) UpdateProject(c *gin.Context) {
 
 }
 
+// DeleteProject godoc
+// @Summary Updates an existing project
+// @Description Updates an existing project
+// @Tags Project
+// @Security ApiKeyAuth
+// @Accept  json
+// @Produce  json
+// @Param   project     body    models.UpdateProjectParams     true        "Project"
+// @Success 204 {string} "ok"
+// @Failure 400 {object} models.Error "Invalid payload"
+// @Failure 500 {object} models.Error "Internal error"
+// @Router /project [put]
 func (ph *ProjectHandler) DeleteProject(c *gin.Context) {
 
 }
