+++ resource-service/handler/service_default_resource_handler.go
@@ -2,7 +2,6 @@ package handler
 
 import (
 	"github.com/gin-gonic/gin"
-	"github.com/keptn/keptn/resource-service/common"
 )
 
 type IServiceDefaultResourceHandler interface {
@@ -16,16 +15,28 @@ type IServiceDefaultResourceHandler interface {
 
 type ServiceDefaultResourceHandler struct {
 	ServiceDefaultResourceManager IServiceDefaultResourceManager
-	EventSender                   common.EventSender
 }
 
-func NewServiceDefaultResourceHandler(serviceDefaultResourceManager IServiceDefaultResourceManager, eventSender common.EventSender) *ServiceDefaultResourceHandler {
+func NewServiceDefaultResourceHandler(serviceDefaultResourceManager IServiceDefaultResourceManager) *ServiceDefaultResourceHandler {
 	return &ServiceDefaultResourceHandler{
 		ServiceDefaultResourceManager: serviceDefaultResourceManager,
-		EventSender:                   eventSender,
 	}
 }
 
+// CreateServiceDefaultResources godoc
+// @Summary Creates service default resource
+// @Description Create list of new default resources used for a service in all stages
+// @Tags Service Default Resource
+// @Security ApiKeyAuth
+// @Accept  json
+// @Produce  json
+// @Param	project				path	string	true	"The name of the project"
+// @Param	service				path	string	true	"The name of the service"
+// @Param   resources     body    models.CreateResourcesParams     true        "List of resources"
+// @Success 204 {string} "ok"
+// @Failure 400 {object} models.Error "Invalid payload"
+// @Failure 500 {object} models.Error "Internal error"
+// @Router /project/{project}/service/{service}/resource [post]
 func (ph *ServiceDefaultResourceHandler) CreateServiceDefaultResources(c *gin.Context) {
 
 }
