+++ resource-service/models/error.go
@@ -1,13 +1,5 @@
 package models
 
-import (
-	"context"
-
-	"github.com/go-openapi/errors"
-	"github.com/go-openapi/strfmt"
-	"github.com/go-openapi/validate"
-)
-
 // Error error
 //
 // swagger:model Error
@@ -19,32 +11,4 @@ type Error struct {
 	// Error message
 	// Required: true
 	Message *string `json:"message"`
-}
-
-// Validate validates this error
-func (m *Error) Validate(formats strfmt.Registry) error {
-	var res []error
-
-	if err := m.validateMessage(formats); err != nil {
-		res = append(res, err)
-	}
-
-	if len(res) > 0 {
-		return errors.CompositeValidationError(res...)
-	}
-	return nil
-}
-
-func (m *Error) validateMessage(formats strfmt.Registry) error {
-
-	if err := validate.Required("message", "body", m.Message); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// ContextValidate validates this error based on context it is used
-func (m *Error) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
-	return nil
 }
