+++ resource-service/models/version.go
@@ -1,11 +1,5 @@
 package models
 
-import (
-	"context"
-
-	"github.com/go-openapi/strfmt"
-)
-
 // Version version
 //
 // swagger:model Version
@@ -19,14 +13,4 @@ type Version struct {
 
 	// version/git commit id of the resource
 	Version string `json:"version,omitempty"`
-}
-
-// Validate validates this version
-func (m *Version) Validate(formats strfmt.Registry) error {
-	return nil
-}
-
-// ContextValidate validates this version based on context it is used
-func (m *Version) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
-	return nil
 }
